requirements.txt
Flask
ultralytics
opencv-python
matplotlib
Pillow
torch
torchvision
numpy
Werkzeug
